+++ CHANGELOG.md
@@ -8,6 +8,9 @@
     * Added new `Grav\Framework\File\Formatter` classes for encoding/decoding YAML, Markdown, JSON, INI and PHP serialized strings
     * Added `Grav\Common\Media` interfaces and trait; use those in `Page` and `Media` classes 
     * Added `Grav\Common\Page` interface to allow custom page types in the future
+    * Added `Grav\Framework\Session` class to replace `RocketTheme\Toolbox\Session\Session`
+1. [](#improved)
+    * Improved session handling, allow all session configuration options in `system.session.options`
 1. [](#bugfix)
     * Fixed bug in `ContentBlock` serialization
     * Fixed `Route::withQueryParam()` to accept array values